Add admin pricing, inventory reporting, and order filter enhancements.
Implement product profit/cost management with auto sale-price recalculation, strengthen warehouse receipt validation and weighted-cost updates, and add inventory lookup/import-export/low-stock reports with supporting SQL indexes.

Made-with: Cursor

+++ admin/san-pham/list.php
@@ -157,7 +157,7 @@ for ($i = 1; $i <= $numberOfPages; $i++) {
                         $index++;
                        $orderNumber = ($currentPage - 1) * $productsPerPage + $index;
                        $unit = isset($value['unit']) ? $value['unit'] : '-';
-                       $cost_price = isset($value['cost_price']) ? (int)$value['cost_price'] : 0;
+                       $cost_price = isset($value['cost_price']) ? (float)$value['cost_price'] : 0;
                        $profit_rate = isset($value['profit_rate']) ? (float)$value['profit_rate'] : 0;
                     ?>
                     <tr>
